+++ src/gql/resolvers/elements/Product.php
@@ -1,0 +1,62 @@
+<?php
+/**
+ * @link https://craftcms.com/
+ * @copyright Copyright (c) Pixel & Tonic, Inc.
+ * @license https://craftcms.github.io/license/
+ */
+
+namespace craft\digitalproducts\gql\resolvers\elements;
+
+use craft\digitalproducts\elements\Product as ProductElement;
+use craft\digitalproducts\helpers\Gql as GqlHelper;
+use craft\gql\base\ElementResolver;
+use craft\helpers\Db;
+
+/**
+ * Class Product
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 2.4
+ */
+class Product extends ElementResolver
+{
+    /**
+     * @inheritdoc
+     */
+    public static function prepareQuery($source, array $arguments, $fieldName = null)
+    {
+        // If this is the beginning of a resolver chain, start fresh
+        if ($source === null) {
+            $query = ProductElement::find();
+            // If not, get the prepared element query
+        } else {
+            $query = $source->$fieldName;
+        }
+
+        // If it's preloaded, it's preloaded.
+        if (is_array($query)) {
+            return $query;
+        }
+
+        foreach ($arguments as $key => $value) {
+            if (method_exists($query, $key)) {
+                $query->$key($value);
+            } elseif (property_exists($query, $key)) {
+                $query->$key = $value;
+            } else {
+                // Catch custom field queries
+                $query->$key($value);
+            }
+        }
+
+        $pairs = GqlHelper::extractAllowedEntitiesFromSchema('read');
+
+        if (!GqlHelper::canQueryProducts()) {
+            return [];
+        }
+
+        $query->andWhere(['in', 'typeId', array_values(Db::idsByUids('{{%digitalproducts_producttypes}}', $pairs['digitalProductTypes']))]);
+
+        return $query;
+    }
+}